requirements.txt
openai
faiss-cpu
numpy
Flask
python-dotenv